# cmd/docs/v2/swagger.yaml
basePath: /
definitions:
  dtos.CreateReservationDTO:
    properties:
      interestedProperty:
        example: 1
        type: integer
      proposedMessage:
        example: I am very interested in leasing this property.
        type: string
      purpose:
        example: Lease agreement for residential property
        type: string
      question:
        example: Could you provide more details about the lease duration?
        type: string
    type: object
...
  dtos.GetReservationDTO:
    properties:
      id:
        example: 1
        type: integer
      interestedProperty:
        example: 1
        type: integer
      lastModified:
        example: "2022-01-01T00:00:00Z"
        type: string
      lesseeID:
        example: 1
        type: integer
      propertyName:
        example: Example Property
        type: string
      proposedMessage:
        example: I am very interested in leasing this property.
        type: string
      purpose:
        example: Lease agreement for residential property
        type: string
      question:
        example: Could you provide more details about the lease duration?
        type: string
      status:
        example: pending
        type: string
    type: object
...
  dtos.UpdateReservationDTO:
    properties:
      proposedMessage:
        example: I would like to update my earlier message.
        type: string
      purpose:
        example: Updated lease purpose
        type: string
      question:
        example: Is there any flexibility in the lease terms?
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      message:
        example: message
        type: string
      status_code:
        example: 888
        type: integer
    type: object
host: localhost:5000/api/v2
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: API documentation for LeaseEase.
  title: LeaseEase API
  version: "2.0"
paths:
...
  /lessee/create:
    post:
      consumes:
      - application/json
      description: Parses the reservation body and creates a new lease reservation
        using the lessee ID from the JWT token.
      parameters:
      - description: Lease Reservation Data
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateReservationDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Reservation created successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Failed to parse reservation body
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create a New Lease Reservation
      tags:
      - Lessee
  /lessee/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a lease reservation using the reservation ID provided in
        the URL.
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reservation deleted successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid reservation ID
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Reservation not found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Delete a Lease Reservation
      tags:
      - Lessee
  /lessee/reservations:
    get:
      consumes:
      - application/json
      description: Retrieves lease reservations for the lessee identified by the JWT
        token.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reservations retrieved successfully
          schema:
            items:
              $ref: '#/definitions/dtos.GetReservationDTO'
            type: array
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get Reservations by Lessee ID
      tags:
      - Lessee
  /lessee/update/{id}:
    put:
      consumes:
      - application/json
      description: Parses the reservation body and updates an existing lease reservation
        identified by its ID.
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lease Reservation Update Data
        in: body
        name: reservation
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateReservationDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Reservation updated successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Failed to parse reservation body or invalid reservation ID
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Reservation not found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update an Existing Lease Reservation
      tags:
      - Lessee
  /lessor/accept/{id}:
    post:
      consumes:
      - application/json
      description: Accept a reservation by ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reservation accepted successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid reservation ID
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Failed to accept reservation
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Accept a reservation
      tags:
      - Lessor
  /lessor/decline/{id}:
    post:
      consumes:
      - application/json
      description: Decline a reservation by ID
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reservation declined successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid reservation ID
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Failed to decline reservation
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Decline a reservation
      tags:
      - Lessor
  /lessor/reservations/{propID}:
    get:
      consumes:
      - application/json
      description: Get reservations by property ID with pagination
      parameters:
      - description: Property ID
        in: path
        name: propID
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reservations retrieved successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid page or pageSize parameter
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Failed to retrieve reservations
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get reservations by property ID
      tags:
      - Lessor
...
swagger: "2.0"